requirements.txt
absl-py==0.9.0
asgiref==3.2.7
astor==0.8.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cloudpickle==1.4.1
decorator==4.4.2
Django==3.0.6
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-rgbfield==1.0
django-tinymce==3.0.2
future==0.18.2
gast==0.2.2
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
idna==2.9
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
music21==5.7.2
numpy==1.18.4
oauthlib==3.1.0
opencv-contrib-python==4.2.0.34
opt-einsum==3.2.1
Pillow==7.1.2
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygan==1.0.8
pylab-sdk==1.1.2
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
sqlparse==0.3.1
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
tensorflow-gan==2.0.0
tensorflow-hub==0.8.0
tensorflow-probability==0.9.0
termcolor==1.1.0
torch==1.5.0
torchvision==0.6.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
django-ckeditor==5.9.0
django-tinymce==3.0.2